[Serve] Expose ApplicationStatus, ApplicationStatusInfo, and StatusOverview classes to API docs
